Moved view restaurant function to homeController to fixe the link in the header to return you to the homepage. Now able to list restaurants with certain cuisines from list page.  Did some formatting to the search page.
