+++ src/main/java/org/launchcode/Liftoff/Project/Restaurant/App/controllers/RestaurantController.java
@@ -22,12 +22,6 @@ public class RestaurantController {
     @Autowired
     private CuisineRepository cuisineRepository;
 
-//    @RequestMapping("")
-//    public String index(Model model) {
-//        model.addAttribute("title", "restaurant");
-//        model.addAttribute("restaurant", restaurantRepository.findAll());
-//        return "index";
-//    }
 
     @GetMapping("")
     public String displayAddRestaurantForm(Model model) {
@@ -49,19 +43,6 @@ public class RestaurantController {
         restaurantRepository.save(newRestaurant);
 
         return "redirect:";
-    }
-
-    @GetMapping("view/{restaurantId}")
-    public String displayViewRestaurant(Model model, @PathVariable int restaurantId) {
-
-        Optional optRestaurant = restaurantRepository.findById(restaurantId);
-        if (optRestaurant.isPresent()) {
-            Restaurant restaurant = (Restaurant) optRestaurant.get();
-            model.addAttribute("restaurant", restaurant);
-            return "view";
-        } else {
-            return "redirect:../";
-        }
     }
 
 }
