+++ src/main/resources/static/css/foodie.css
@@ -89,4 +89,16 @@ form {
   margin-top: 10px;
 }
 
+.search {
+    margin-left: 50px;
+    margin-top: 25px;
+}
+
+.list {
+    margin-left: auto;
+    margin-right: auto;
+    width: 700px;
+    text-align: center;
+}
+
