+++ src/main/resources/templates/fragments.html
@@ -5,6 +5,7 @@
     <title>Foodie</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.0/dist/css/bootstrap.min.css" rel="stylesheet"/>
     <link href='https://fonts.googleapis.com/css?family=Oleo Script Swash Caps' rel='stylesheet'/>
+    <link href='https://fonts.googleapis.com/css?family=Belgrano' rel='stylesheet'/>
     <link th:href="@{/css/foodie.css}" rel="stylesheet" />
     <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.0/dist/js/bootstrap.min.js"></script>
