+++ src/main/resources/templates/view.html
@@ -8,7 +8,7 @@
 
     <div class="row">
         <div class="column">
-            <h1 class="restaurantName" th:text="${restaurant.name}" >Default Title</h1>
+            <h1 class="restaurantName" th:text="${restaurant.name}" ></h1>
             <img class="largeImage" src="/images/texasRoadhouse.jpg" width="600" height="300">
 
             <div>
@@ -38,6 +38,10 @@
             </div>
 
         </div>
+    </div>
+
+    <div class="text-center">
+        <div th:replace="fragments :: app-name-auth"></div>
     </div>
 
 </body>
